+++ lib/paper_trail_diff/time_version_range.rb
@@ -42,7 +42,7 @@ module PaperTrailDiff
 
     #: (untyped) -> untyped
     def trailing_version(relation)
-      @time_range.trailing_scope(relation).order(created_at: :asc, id: :asc).first
+      @time_range.trailing_scope(relation).reorder(created_at: :asc, id: :asc).first
     end
 
     #: (Array[untyped]) -> Array[untyped]